Created new Porjects and seprated Validationa and NHL Service. Also refactored and cleaned structure
