--- Services/Sql/Interface/IDraftSql.cs
@@ -1,10 +1,0 @@
-﻿using System.Data;
-using Tietoa.Domain.Models.Draft;
-
-namespace Services.Sql.Interface
-{
-    public interface IDraftSql
-    {
-        DataTable InsertDraftTable(int ProspectId, string Round, int Pick, string Team, string FullName, int DraftYear);
-    }
-}
