--- Services/Sql/Interface/IGetData.cs
@@ -1,10 +1,0 @@
-﻿//using System.Data;
-
-//namespace Services.Sql.Interface
-
-//{
-//    public interface IGetData
-//    {
-//        DataTable GetDataFromTable(string TableName);
-//    }
-//}
