--- Services/Sql/Interface/IInsertData.cs
@@ -1,10 +1,0 @@
-﻿using System.Data;
-
-namespace Services.Sql.Interface
-
-{
-    public interface IInsertData
-    {
-        DataTable InsertTable(int divisionId, string divisionName, string divisionConference);
-    }
-}
