+++ Tietoa.sln
@@ -4,10 +4,24 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.2.32630.192
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tietoa", "Tietoa\Tietoa.csproj", "{DEDCDE21-0AEB-4EC3-B6DA-2E89524EC035}"
+	ProjectSection(ProjectDependencies) = postProject
+		{3241E17A-4937-49FD-8FD3-C478FB581CD9} = {3241E17A-4937-49FD-8FD3-C478FB581CD9}
+	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Services", "Services\Services.csproj", "{77AEF13F-E6B4-47B3-8C00-7E59196FB476}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tietoa.Domain", "Tietoa.Domain\Tietoa.Domain.csproj", "{C1284A30-6610-4FCD-BF0B-290EB7DC5BA4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tietoa.Validation", "Tietoa.Validation\Tietoa.Validation.csproj", "{06D03D6D-A44D-4272-BE29-6FFDBD2385A1}"
+	ProjectSection(ProjectDependencies) = postProject
+		{3241E17A-4937-49FD-8FD3-C478FB581CD9} = {3241E17A-4937-49FD-8FD3-C478FB581CD9}
+		{C1284A30-6610-4FCD-BF0B-290EB7DC5BA4} = {C1284A30-6610-4FCD-BF0B-290EB7DC5BA4}
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tietoa.Services", "Tietoa.Services\Tietoa.Services.csproj", "{3241E17A-4937-49FD-8FD3-C478FB581CD9}"
+	ProjectSection(ProjectDependencies) = postProject
+		{C1284A30-6610-4FCD-BF0B-290EB7DC5BA4} = {C1284A30-6610-4FCD-BF0B-290EB7DC5BA4}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +41,14 @@ Global
 		{C1284A30-6610-4FCD-BF0B-290EB7DC5BA4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C1284A30-6610-4FCD-BF0B-290EB7DC5BA4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C1284A30-6610-4FCD-BF0B-290EB7DC5BA4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{06D03D6D-A44D-4272-BE29-6FFDBD2385A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06D03D6D-A44D-4272-BE29-6FFDBD2385A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06D03D6D-A44D-4272-BE29-6FFDBD2385A1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06D03D6D-A44D-4272-BE29-6FFDBD2385A1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3241E17A-4937-49FD-8FD3-C478FB581CD9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3241E17A-4937-49FD-8FD3-C478FB581CD9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3241E17A-4937-49FD-8FD3-C478FB581CD9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3241E17A-4937-49FD-8FD3-C478FB581CD9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
