+++ Tietoa/Tietoa.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="Newtonsoft.Json.Bson" Version="1.0.2" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
     <PackageReference Include="System.Runtime.Serialization.Json" Version="4.3.0" />
@@ -16,6 +16,8 @@
   <ItemGroup>
     <ProjectReference Include="..\Services\Services.csproj" />
     <ProjectReference Include="..\Tietoa.Domain\Tietoa.Domain.csproj" />
+    <ProjectReference Include="..\Tietoa.Services\Tietoa.Services.csproj" />
+    <ProjectReference Include="..\Tietoa.Validation\Tietoa.Validation.csproj" />
   </ItemGroup>
 
 </Project>
